[profiler] move hb_profiler to hammerblade/profiler
