+++ torch/hammerblade/profiler/__init__.py
@@ -20,7 +20,7 @@ def is_in_ROI():
     return profiler_status.is_in_ROI
 
 # --------- import components ---------
-import torch.hb_profiler.exec_time
-import torch.hb_profiler.unimpl
-import torch.hb_profiler.chart
-import torch.hb_profiler.route
+import torch.hammerblade.profiler.exec_time
+import torch.hammerblade.profiler.unimpl
+import torch.hammerblade.profiler.chart
+import torch.hammerblade.profiler.route
